Extract and document GPU command graph compiler
