+++ modules/experimental/src/gpu-primitives/gpu-compaction.ts
@@ -19,23 +19,45 @@ const COMPACTION_WORKGROUP_SIZE = 256;
 /** Packed uint32 graph data accepted by {@link GPUCompaction}. */
 export type GPUCompactionInput = GraphDataView<'uint32'> | GraphVectorView<'uint32'>;
 
-/** Properties for stable graph-native selection by uint32 flags. */
+/** Properties for one graph-native stable compaction. */
 export type GPUCompactionProps = {
+  /** Prefix for generated graph node and transient resource IDs. */
   id?: string;
+  /** Packed source values as one data view or an ordered vector. */
   input: GPUCompactionInput;
+  /** Matching packed selection flags; zero rejects and any non-zero value accepts. */
   flags: GPUCompactionInput;
+  /** Caller-owned destination with matching view kind and sufficient capacity or topology. */
   output: GPUCompactionInput;
+  /** Caller-owned view whose first row receives the accepted value count. */
   count: GraphDataView<'uint32'>;
 };
 
-/** Stable graph-composed compaction of uint32 values selected by 0/1 flags. */
+/**
+ * Stable graph-composed compaction of `uint32` values selected by flags.
+ *
+ * An exclusive {@link GPUScan} converts flags to output offsets. A scatter pass then preserves
+ * source order while writing accepted values and the final accepted count. Vector inputs are one
+ * logical sequence whose selected values fill the existing output topology.
+ */
 export class GPUCompaction {
+  /** Prefix for generated graph node and transient resource IDs. */
   readonly id: string;
+  /** Packed source values or ordered source vector. */
   readonly input: GPUCompactionInput;
+  /** Packed selection flags with the same view kind and topology as the input. */
   readonly flags: GPUCompactionInput;
+  /** Caller-owned compacted destination with matching view kind. */
   readonly output: GPUCompactionInput;
+  /** Caller-owned accepted-count destination. */
   readonly count: GraphDataView<'uint32'>;
 
+  /**
+   * Creates and validates a stable compaction description.
+   *
+   * @throws If views are not packed `uint32` data, input/flag lengths differ, or an output lacks
+   * capacity.
+   */
   constructor(props: GPUCompactionProps) {
     this.id = props.id ?? 'gpu-compaction';
     this.input = props.input;
@@ -69,7 +91,12 @@ export class GPUCompaction {
     }
   }
 
-  /** Adds scan and scatter passes to the target graph. */
+  /**
+   * Adds scan and scatter passes to the target graph.
+   *
+   * Empty input adds only a pass that writes a zero count. This method does not submit or read back
+   * commands.
+   */
   addToGraph<Parameters>(graph: GPUCommandGraph<Parameters>): void {
     for (const view of [
       ...getCompactionChunks(this.input),
@@ -87,72 +114,16 @@ export class GPUCompaction {
       return;
     }
 
-    if (
-      this.input instanceof GraphVectorView &&
-      this.flags instanceof GraphVectorView &&
-      this.output instanceof GraphVectorView
-    ) {
-      addVectorCompaction(graph, this.id, this.input, this.flags, this.output, this.count);
-    } else {
-      const input = this.input as GraphDataView<'uint32'>;
-      const flags = this.flags as GraphDataView<'uint32'>;
-      const output = this.output as GraphDataView<'uint32'>;
-      const offsets = createTransientView(graph, `${this.id}-offsets`, 'uint32', input.length);
-      new GPUScan({id: `${this.id}-scan`, input: flags, output: offsets}).addToGraph(graph);
-      addScatterPass(graph, {
-        id: `${this.id}-scatter`,
-        input,
-        flags,
-        offsets,
-        output,
-        count: this.count
-      });
-    }
+    const offsets =
+      this.flags instanceof GraphVectorView
+        ? createTransientVectorView(graph, `${this.id}-offsets`, this.flags)
+        : createTransientView(graph, `${this.id}-offsets`, 'uint32', this.flags.length);
+    new GPUScan({id: `${this.id}-scan`, input: this.flags, output: offsets}).addToGraph(graph);
+    addScatterPasses(graph, this.id, this.input, this.flags, offsets, this.output, this.count);
   }
 }
 
-function addVectorCompaction<Parameters>(
-  graph: GPUCommandGraph<Parameters>,
-  id: string,
-  input: GraphVectorView<'uint32'>,
-  flags: GraphVectorView<'uint32'>,
-  output: GraphVectorView<'uint32'>,
-  count: GraphDataView<'uint32'>
-): void {
-  const offsets = createTransientVectorView(graph, `${id}-offsets`, flags);
-  new GPUScan({id: `${id}-scan`, input: flags, output: offsets}).addToGraph(graph);
-
-  let outputStart = 0;
-  for (let outputChunkIndex = 0; outputChunkIndex < output.data.length; outputChunkIndex++) {
-    const outputChunk = output.data[outputChunkIndex];
-    const outputEnd = outputStart + outputChunk.length;
-    if (outputChunk.length > 0) {
-      for (let inputChunkIndex = 0; inputChunkIndex < input.data.length; inputChunkIndex++) {
-        if (input.data[inputChunkIndex].length > 0) {
-          addVectorScatterPass(graph, {
-            id: `${id}-scatter-input-${inputChunkIndex}-output-${outputChunkIndex}`,
-            input: input.data[inputChunkIndex],
-            flags: flags.data[inputChunkIndex],
-            offsets: offsets.data[inputChunkIndex],
-            output: outputChunk,
-            outputStart,
-            outputEnd
-          });
-        }
-      }
-    }
-    outputStart = outputEnd;
-  }
-
-  const lastChunkIndex = findLastNonEmptyChunk(input);
-  addCountPass(graph, {
-    id: `${id}-write-count`,
-    flags: flags.data[lastChunkIndex],
-    offsets: offsets.data[lastChunkIndex],
-    count
-  });
-}
-
+/** Writes the required zero count for an empty input. */
 function addClearCountPass<Parameters>(
   graph: GPUCommandGraph<Parameters>,
   id: string,
@@ -183,7 +154,58 @@ function addClearCountPass<Parameters>(
   });
 }
 
-function addVectorScatterPass<Parameters>(
+/** Routes every non-empty input chunk through each non-empty logical output range. */
+function addScatterPasses<Parameters>(
+  graph: GPUCommandGraph<Parameters>,
+  id: string,
+  input: GPUCompactionInput,
+  flags: GPUCompactionInput,
+  offsets: GPUCompactionInput,
+  output: GPUCompactionInput,
+  count: GraphDataView<'uint32'>
+): void {
+  const inputChunks = getCompactionChunks(input);
+  const flagChunks = getCompactionChunks(flags);
+  const offsetChunks = getCompactionChunks(offsets);
+  const outputChunks = getCompactionChunks(output);
+  const inputChunkIndices = inputChunks
+    .map((chunk, chunkIndex) => (chunk.length > 0 ? chunkIndex : -1))
+    .filter(chunkIndex => chunkIndex >= 0);
+  const outputChunkIndices = outputChunks
+    .map((chunk, chunkIndex) => (chunk.length > 0 ? chunkIndex : -1))
+    .filter(chunkIndex => chunkIndex >= 0);
+  const countInputChunkIndex = inputChunkIndices[inputChunkIndices.length - 1];
+  const countOutputChunkIndex = outputChunkIndices[outputChunkIndices.length - 1];
+  const isVector = input instanceof GraphVectorView;
+
+  let outputStart = 0;
+  for (let outputChunkIndex = 0; outputChunkIndex < outputChunks.length; outputChunkIndex++) {
+    const outputChunk = outputChunks[outputChunkIndex];
+    const outputEnd = outputStart + outputChunk.length;
+    if (outputChunk.length > 0) {
+      for (const inputChunkIndex of inputChunkIndices) {
+        const writesCount =
+          inputChunkIndex === countInputChunkIndex && outputChunkIndex === countOutputChunkIndex;
+        addScatterPass(graph, {
+          id: isVector
+            ? `${id}-scatter-input-${inputChunkIndex}-output-${outputChunkIndex}`
+            : `${id}-scatter`,
+          input: inputChunks[inputChunkIndex],
+          flags: flagChunks[inputChunkIndex],
+          offsets: offsetChunks[inputChunkIndex],
+          output: outputChunk,
+          outputStart,
+          outputEnd,
+          count: writesCount ? count : undefined
+        });
+      }
+    }
+    outputStart = outputEnd;
+  }
+}
+
+/** Scatters one input chunk into one logical output range and optionally writes the total count. */
+function addScatterPass<Parameters>(
   graph: GPUCommandGraph<Parameters>,
   props: {
     id: string;
@@ -193,8 +215,17 @@ function addVectorScatterPass<Parameters>(
     output: GraphDataView<'uint32'>;
     outputStart: number;
     outputEnd: number;
+    count?: GraphDataView<'uint32'>;
   }
 ): void {
+  const countBinding = props.count
+    ? '@group(0) @binding(4) var<storage, read_write> outputCount: array<u32>;'
+    : '';
+  const countWrite = props.count
+    ? `if (index == ELEMENT_COUNT - 1u) {
+    outputCount[COUNT_OFFSET] = outputIndex + flag;
+  }`
+    : '';
   const source = /* wgsl */ `
 const ELEMENT_COUNT: u32 = ${props.input.length}u;
 const INPUT_OFFSET: u32 = ${getViewElementOffset(props.input)}u;
@@ -203,20 +234,24 @@ const OFFSETS_OFFSET: u32 = ${getViewElementOffset(props.offsets)}u;
 const OUTPUT_OFFSET: u32 = ${getViewElementOffset(props.output)}u;
 const OUTPUT_START: u32 = ${props.outputStart}u;
 const OUTPUT_END: u32 = ${props.outputEnd}u;
+${props.count ? `const COUNT_OFFSET: u32 = ${getViewElementOffset(props.count)}u;` : ''}
 @group(0) @binding(0) var<storage, read> inputValues: array<u32>;
 @group(0) @binding(1) var<storage, read> flags: array<u32>;
 @group(0) @binding(2) var<storage, read> offsets: array<u32>;
 @group(0) @binding(3) var<storage, read_write> outputValues: array<u32>;
+${countBinding}
 
 @compute @workgroup_size(${COMPACTION_WORKGROUP_SIZE}) fn main(
   @builtin(global_invocation_id) globalId: vec3<u32>
 ) {
   let index = globalId.x;
-  if (index >= ELEMENT_COUNT || min(flags[FLAGS_OFFSET + index], 1u) == 0u) { return; }
+  if (index >= ELEMENT_COUNT) { return; }
+  let flag = min(flags[FLAGS_OFFSET + index], 1u);
   let outputIndex = offsets[OFFSETS_OFFSET + index];
-  if (outputIndex >= OUTPUT_START && outputIndex < OUTPUT_END) {
+  if (flag != 0u && outputIndex >= OUTPUT_START && outputIndex < OUTPUT_END) {
     outputValues[OUTPUT_OFFSET + outputIndex - OUTPUT_START] = inputValues[INPUT_OFFSET + index];
   }
+  ${countWrite}
 }`;
   addCompactionPass(graph, {
     id: props.id,
@@ -225,52 +260,21 @@ const OUTPUT_END: u32 = ${props.outputEnd}u;
       {buffer: props.input, usage: 'storage-read'},
       {buffer: props.flags, usage: 'storage-read'},
       {buffer: props.offsets, usage: 'storage-read'},
-      {buffer: props.output, usage: 'storage-write'}
+      {buffer: props.output, usage: 'storage-write'},
+      ...(props.count ? [{buffer: props.count, usage: 'storage-write'} as const] : [])
     ],
     bindings: {
       inputValues: props.input,
       flags: props.flags,
       offsets: props.offsets,
-      outputValues: props.output
+      outputValues: props.output,
+      ...(props.count ? {outputCount: props.count} : {})
     },
     dispatchCount: Math.ceil(props.input.length / COMPACTION_WORKGROUP_SIZE)
   });
 }
 
-function addCountPass<Parameters>(
-  graph: GPUCommandGraph<Parameters>,
-  props: {
-    id: string;
-    flags: GraphDataView<'uint32'>;
-    offsets: GraphDataView<'uint32'>;
-    count: GraphDataView<'uint32'>;
-  }
-): void {
-  const lastIndex = props.flags.length - 1;
-  const source = /* wgsl */ `
-const FLAGS_OFFSET: u32 = ${getViewElementOffset(props.flags)}u;
-const OFFSETS_OFFSET: u32 = ${getViewElementOffset(props.offsets)}u;
-const COUNT_OFFSET: u32 = ${getViewElementOffset(props.count)}u;
-@group(0) @binding(0) var<storage, read> flags: array<u32>;
-@group(0) @binding(1) var<storage, read> offsets: array<u32>;
-@group(0) @binding(2) var<storage, read_write> outputCount: array<u32>;
-@compute @workgroup_size(1) fn main() {
-  outputCount[COUNT_OFFSET] = offsets[OFFSETS_OFFSET + ${lastIndex}u] +
-    min(flags[FLAGS_OFFSET + ${lastIndex}u], 1u);
-}`;
-  addCompactionPass(graph, {
-    id: props.id,
-    source,
-    resources: [
-      {buffer: props.flags, usage: 'storage-read'},
-      {buffer: props.offsets, usage: 'storage-read'},
-      {buffer: props.count, usage: 'storage-write'}
-    ],
-    bindings: {flags: props.flags, offsets: props.offsets, outputCount: props.count},
-    dispatchCount: 1
-  });
-}
-
+/** Adds a storage-only computation pass used by compaction kernels. */
 function addCompactionPass<Parameters>(
   graph: GPUCommandGraph<Parameters>,
   props: {
@@ -315,99 +319,14 @@ function addCompactionPass<Parameters>(
   });
 }
 
-function addScatterPass<Parameters>(
-  graph: GPUCommandGraph<Parameters>,
-  props: {
-    id: string;
-    input: GraphDataView<'uint32'>;
-    flags: GraphDataView<'uint32'>;
-    offsets: GraphDataView<'uint32'>;
-    output: GraphDataView<'uint32'>;
-    count: GraphDataView<'uint32'>;
-  }
-): void {
-  const length = props.input.length;
-  const source = /* wgsl */ `
-const ELEMENT_COUNT: u32 = ${length}u;
-const INPUT_OFFSET: u32 = ${getViewElementOffset(props.input)}u;
-const FLAGS_OFFSET: u32 = ${getViewElementOffset(props.flags)}u;
-const OFFSETS_OFFSET: u32 = ${getViewElementOffset(props.offsets)}u;
-const OUTPUT_OFFSET: u32 = ${getViewElementOffset(props.output)}u;
-const COUNT_OFFSET: u32 = ${getViewElementOffset(props.count)}u;
-@group(0) @binding(0) var<storage, read> inputValues: array<u32>;
-@group(0) @binding(1) var<storage, read> flags: array<u32>;
-@group(0) @binding(2) var<storage, read> offsets: array<u32>;
-@group(0) @binding(3) var<storage, read_write> outputValues: array<u32>;
-@group(0) @binding(4) var<storage, read_write> outputCount: array<u32>;
-
-@compute @workgroup_size(${COMPACTION_WORKGROUP_SIZE}) fn main(
-  @builtin(global_invocation_id) globalId: vec3<u32>
-) {
-  let index = globalId.x;
-  if (index >= ELEMENT_COUNT) { return; }
-  let flag = min(flags[FLAGS_OFFSET + index], 1u);
-  if (flag != 0u) {
-    outputValues[OUTPUT_OFFSET + offsets[OFFSETS_OFFSET + index]] = inputValues[INPUT_OFFSET + index];
-  }
-  if (index == ELEMENT_COUNT - 1u) {
-    outputCount[COUNT_OFFSET] = offsets[OFFSETS_OFFSET + index] + flag;
-  }
-}`;
-  graph.addComputePass({
-    id: props.id,
-    resources: [
-      {buffer: props.input, usage: 'storage-read'},
-      {buffer: props.flags, usage: 'storage-read'},
-      {buffer: props.offsets, usage: 'storage-read'},
-      {buffer: props.output, usage: 'storage-write'},
-      {buffer: props.count, usage: 'storage-write'}
-    ],
-    compile: ({device}) => {
-      const computation = new Computation(device, {
-        id: props.id,
-        source,
-        shaderLayout: {
-          bindings: [
-            {name: 'inputValues', type: 'storage', group: 0, location: 0},
-            {name: 'flags', type: 'storage', group: 0, location: 1},
-            {name: 'offsets', type: 'storage', group: 0, location: 2},
-            {name: 'outputValues', type: 'storage', group: 0, location: 3},
-            {name: 'outputCount', type: 'storage', group: 0, location: 4}
-          ]
-        }
-      });
-      return {
-        encode: ({computePass, getBuffer}) => {
-          computation.setBindings({
-            inputValues: getViewBinding(props.input, getBuffer),
-            flags: getViewBinding(props.flags, getBuffer),
-            offsets: getViewBinding(props.offsets, getBuffer),
-            outputValues: getViewBinding(props.output, getBuffer),
-            outputCount: getViewBinding(props.count, getBuffer)
-          });
-          computation.dispatch(computePass, Math.ceil(length / COMPACTION_WORKGROUP_SIZE));
-        },
-        destroy: () => computation.destroy()
-      };
-    }
-  });
-}
-
+/** Normalizes one compaction input or vector into its ordered atomic chunks. */
 function getCompactionChunks(input: GPUCompactionInput): readonly GraphDataView<'uint32'>[] {
   return input instanceof GraphVectorView ? input.data : [input];
 }
 
+/** Validates every atomic chunk accepted by a compaction input. */
 function validateCompactionInput(input: GPUCompactionInput, name: string): void {
   for (const chunk of getCompactionChunks(input)) {
     validatePackedUint32View(chunk, name);
   }
-}
-
-function findLastNonEmptyChunk(input: GraphVectorView<'uint32'>): number {
-  for (let chunkIndex = input.data.length - 1; chunkIndex >= 0; chunkIndex--) {
-    if (input.data[chunkIndex].length > 0) {
-      return chunkIndex;
-    }
-  }
-  throw new Error('GPUCompaction requires a non-empty chunk');
 }
